Introduce abstract classes
