+++ scm-webapp/src/main/java/sonia/scm/repository/cli/NamespacePermissionsAddCommand.java
@@ -28,82 +28,31 @@ import com.google.common.annotations.VisibleForTesting;
 import picocli.CommandLine;
 import sonia.scm.cli.ParentCommand;
 import sonia.scm.cli.PermissionDescriptionResolver;
+import sonia.scm.repository.Namespace;
 import sonia.scm.repository.NamespaceManager;
-import sonia.scm.repository.RepositoryPermission;
 import sonia.scm.repository.RepositoryRoleManager;
 
 import javax.inject.Inject;
-import java.util.Arrays;
-import java.util.Set;
-
-import static java.util.Arrays.asList;
 
 @CommandLine.Command(name = "add-permissions")
 @ParentCommand(value = NamespaceCommand.class)
-class NamespacePermissionsAddCommand extends NamespacePermissionBaseCommand implements Runnable {
-
-  private final PermissionDescriptionResolver permissionDescriptionResolver;
+class NamespacePermissionsAddCommand extends PermissionsAddCommand<Namespace> implements Runnable {
 
   @CommandLine.Parameters(paramLabel = "namespace", index = "0", descriptionKey = "scm.namespace.add-permissions.namespace")
   private String namespace;
-  @CommandLine.Parameters(paramLabel = "name", index = "1", descriptionKey = "scm.namespace.add-permissions.name")
-  private String name;
-  @CommandLine.Parameters(paramLabel = "verbs", index = "2..", arity = "1..", descriptionKey = "scm.namespace.add-permissions.verbs")
-  private String[] verbs = new String[0];
-
-  @CommandLine.Option(names = {"--group", "-g"}, descriptionKey = "scm.namespace.add-permissions.forGroup")
-  private boolean forGroup;
 
   @Inject
-  NamespacePermissionsAddCommand(NamespaceManager namespaceManager, RepositoryPermissionBaseCommand permissionCommandManager, RepositoryRoleManager roleManager, PermissionDescriptionResolver permissionDescriptionResolver, RepositoryTemplateRenderer templateRenderer) {
-    super(namespaceManager, roleManager, templateRenderer);
-    this.permissionDescriptionResolver = permissionDescriptionResolver;
+  NamespacePermissionsAddCommand(NamespaceManager namespaceManager, RepositoryRoleManager roleManager, PermissionDescriptionResolver permissionDescriptionResolver, RepositoryTemplateRenderer templateRenderer) {
+    super(roleManager, permissionDescriptionResolver, templateRenderer, new NamespacePermissionBaseAdapter(namespaceManager, templateRenderer));
   }
 
   @Override
-  public void run() {
-    modify(
-      namespace,
-      ns -> {
-        if (!Arrays.stream(verbs).allMatch(this::verifyVerbExists)) {
-          return false;
-        }
-        Set<String> resultingVerbs =
-          getPermissionsAsModifiableSet(ns, name, forGroup);
-        if (resultingVerbs.containsAll(asList(this.verbs))) {
-          return false;
-        }
-        resultingVerbs.addAll(asList(this.verbs));
-        replacePermission(ns, new RepositoryPermission(name, resultingVerbs, forGroup));
-        return true;
-      }
-    );
-  }
-
-  private boolean verifyVerbExists(String verb) {
-    if (permissionDescriptionResolver.getDescription(verb).isEmpty()) {
-      renderVerbNotFoundError();
-      return false;
-    }
-    return true;
+  String getIdentifier() {
+    return namespace;
   }
 
   @VisibleForTesting
   void setNamespace(String namespace) {
     this.namespace = namespace;
-  }
-
-  @VisibleForTesting
-  void setName(String name) {
-    this.name = name;
-  }
-
-  @VisibleForTesting
-  void setVerbs(String... verbs) {
-    this.verbs = verbs;
-  }
-
-  public void setForGroup(boolean forGroup) {
-    this.forGroup = forGroup;
   }
 }
