+++ scm-webapp/src/main/java/sonia/scm/repository/cli/NamespacePermissionsRemoveCommand.java
@@ -27,66 +27,30 @@ package sonia.scm.repository.cli;
 import com.google.common.annotations.VisibleForTesting;
 import picocli.CommandLine;
 import sonia.scm.cli.ParentCommand;
+import sonia.scm.repository.Namespace;
 import sonia.scm.repository.NamespaceManager;
-import sonia.scm.repository.RepositoryPermission;
 import sonia.scm.repository.RepositoryRoleManager;
 
 import javax.inject.Inject;
-import java.util.Set;
-
-import static java.util.Arrays.asList;
 
 @CommandLine.Command(name = "remove-permissions")
 @ParentCommand(value = NamespaceCommand.class)
-class NamespacePermissionsRemoveCommand extends NamespacePermissionBaseCommand implements Runnable {
+class NamespacePermissionsRemoveCommand extends PermissionsRemoveCommand<Namespace> {
 
   @CommandLine.Parameters(paramLabel = "namespace", index = "0", descriptionKey = "scm.namespace.remove-permissions.namespace")
   private String namespace;
-  @CommandLine.Parameters(paramLabel = "name", index = "1", descriptionKey = "scm.namespace.remove-permissions.name")
-  private String name;
-  @CommandLine.Parameters(paramLabel = "verbs", index = "2..", arity = "1..", descriptionKey = "scm.namespace.remove-permissions.verbs")
-  private String[] verbs = new String[0];
-
-  @CommandLine.Option(names = {"--group", "-g"}, descriptionKey = "scm.repo.remove-permissions.forGroup")
-  private boolean forGroup;
-
   @Inject
   public NamespacePermissionsRemoveCommand(NamespaceManager namespaceManager, RepositoryRoleManager roleManager, RepositoryTemplateRenderer templateRenderer) {
-    super(namespaceManager, roleManager, templateRenderer);
+    super(roleManager, templateRenderer, new NamespacePermissionBaseAdapter(namespaceManager, templateRenderer));
   }
 
   @Override
-  public void run() {
-    modify(
-      namespace,
-      ns -> {
-        Set<String> resultingVerbs =
-          getPermissionsAsModifiableSet(ns, name, forGroup);
-        if (resultingVerbs.stream().noneMatch(verb -> asList(verbs).contains(verb))) {
-          return false;
-        }
-        resultingVerbs.removeAll(asList(this.verbs));
-        replacePermission(ns, new RepositoryPermission(name, resultingVerbs, forGroup));
-        return true;
-      }
-    );
+  String getIdentifier() {
+    return namespace;
   }
 
   @VisibleForTesting
   void setNamespace(String namespace) {
     this.namespace = namespace;
-  }
-
-  @VisibleForTesting
-  void setName(String name) {
-    this.name = name;
-  }
-
-  public void setVerbs(String... verbs) {
-    this.verbs = verbs;
-  }
-
-  public void setForGroup(boolean forGroup) {
-    this.forGroup = forGroup;
   }
 }
