+++ scm-webapp/src/main/java/sonia/scm/repository/cli/NamespacePermissionsSetRoleCommand.java
@@ -27,65 +27,31 @@ package sonia.scm.repository.cli;
 import com.google.common.annotations.VisibleForTesting;
 import picocli.CommandLine;
 import sonia.scm.cli.ParentCommand;
+import sonia.scm.repository.Namespace;
 import sonia.scm.repository.NamespaceManager;
-import sonia.scm.repository.RepositoryPermission;
 import sonia.scm.repository.RepositoryRoleManager;
 
 import javax.inject.Inject;
 
 @CommandLine.Command(name = "set-role")
 @ParentCommand(value = NamespaceCommand.class)
-class NamespacePermissionsSetRoleCommand extends NamespacePermissionBaseCommand implements Runnable {
-
-  private final RepositoryRoleManager roleManager;
+class NamespacePermissionsSetRoleCommand extends PermissionsSetRoleCommand<Namespace> {
 
   @CommandLine.Parameters(paramLabel = "namespace", index = "0", descriptionKey = "scm.namespace.set-role.namespace")
   private String namespace;
-  @CommandLine.Parameters(paramLabel = "name", index = "1", descriptionKey = "scm.namespace.set-role.name")
-  private String name;
-  @CommandLine.Parameters(paramLabel = "role", index = "2", descriptionKey = "scm.namespace.set-role.role")
-  private String role;
-
-  @CommandLine.Option(names = {"--group", "-g"}, descriptionKey = "scm.namespace.set-role.forGroup")
-  private boolean forGroup;
 
   @Inject
   public NamespacePermissionsSetRoleCommand(NamespaceManager namespaceManager, RepositoryRoleManager roleManager, RepositoryTemplateRenderer templateRenderer) {
-    super(namespaceManager, roleManager, templateRenderer);
-    this.roleManager = roleManager;
+    super(roleManager, templateRenderer, new NamespacePermissionBaseAdapter(namespaceManager, templateRenderer));
   }
 
   @Override
-  public void run() {
-    modify(
-      namespace,
-      ns -> {
-        if (roleManager.get(role) == null) {
-          renderRoleNotFoundError();
-          return false;
-        }
-        replacePermission(ns, new RepositoryPermission(name, role, forGroup));
-        return true;
-      }
-    );
+  protected String getIdentifier() {
+    return namespace;
   }
 
   @VisibleForTesting
   void setNamespace(String namespace) {
     this.namespace = namespace;
-  }
-
-  @VisibleForTesting
-  void setName(String name) {
-    this.name = name;
-  }
-
-  @VisibleForTesting
-  void setRole(String role) {
-    this.role = role;
-  }
-
-  public void setForGroup(boolean forGroup) {
-    this.forGroup = forGroup;
   }
 }
