+++ scm-webapp/src/main/resources/sonia/scm/cli/i18n.properties
@@ -147,31 +147,18 @@ scm.namespace.available-permissions.usage.description.0 = List available permiss
 
 scm.namespace.list-permissions.usage.description.0 = List permissions for namespace
 scm.namespace.list-permissions.namespace = Namespace
-scm.namespace.list-permissions.verbose = Show single detailed permissions
-scm.namespace.list-permissions.keys = Use keys instead of descriptive permission
 
 scm.namespace.add-permissions.usage.description.0 = Add single permissions for user or group
 scm.namespace.add-permissions.namespace = Namespace
-scm.namespace.add-permissions.name = User or group name to grant permission to
-scm.namespace.add-permissions.verbs = Single permissions to grant to user or group
-scm.namespace.add-permissions.forGroup = Add permission for a group, not for a user
 
 scm.namespace.set-role.usage.description.0 = Set role permission for user or group
 scm.namespace.set-role.namespace = Namespace
-scm.namespace.set-role.name = User or group name to set role for
-scm.namespace.set-role.role = Permission role to grant to user or group
-scm.namespace.set-role.forGroup = Set role for a group, not for a user
 
 scm.namespace.remove-permissions.usage.description.0 = Revoke single permission from user or group
 scm.namespace.remove-permissions.namespace = Namespace
-scm.namespace.remove-permissions.name = User or group name to revoke permission from
-scm.namespace.remove-permissions.verbs = Single permissions to revoke from user or group
-scm.namespace.remove-permissions.forGroup = Revoke permission from a group, not for a user
 
 scm.namespace.clear-permissions.usage.description.0 = Revoke all permissions from user or group
 scm.namespace.clear-permissions.namespace = Namespace
-scm.namespace.clear-permissions.name = User or group name to revoke permissions from
-scm.namespace.clear-permissions.forGroup = Revoke permissions from a group, not for a user
 
 ## User
 scm.user.username = Username
