+++ scm-webapp/src/main/resources/sonia/scm/cli/i18n_de.properties
@@ -143,16 +143,22 @@ verbNotFound=  konnte nicht gefunden werden
 scm.namespace.usage.description.0 = Ressourcenkommando für alle Aktionen zu Namespaces
 scm.namespace.list.usage.description.0 = Listet alle Namespaces
 
+scm.namespace.available-permissions.usage.description.0 = Listet verfügbare Berechtigungen für den Namespace
+
 scm.namespace.list-permissions.usage.description.0 = Listet die Berechtigungen für den Namespace
-scm.namespace.list-permissions.repository = Namespace
-scm.namespace.list-permissions.verbose = Zeigt detaillierte einzelne Berechtigungen
-scm.namespace.list-permissions.keys = Nutze Schlüssel anstelle der Beschreibungen
+scm.namespace.list-permissions.namespace = Namespace
 
 scm.namespace.add-permissions.usage.description.0 = Fügt einzelne Berechtigungen für einen Benutzer oder eine Gruppe hinzu
-scm.namespace.add-permissions.repository = Namespace
-scm.namespace.add-permissions.name = Name des Benutzers oder der Gruppe für die Berechtigung
-scm.namespace.add-permissions.verbs = Einzelne Berechtigungen, die für den Benutzer oder die Gruppe hinzugefügt werden sollen
-scm.namespace.add-permissions.forGroup = Setzt die Berechtigung für eine Gruppe, nicht für einen Benutzer
+scm.namespace.add-permissions.namespace = Namespace
+
+scm.namespace.set-role.usage.description.0 = Setzt eine Berechtigungesrolle für einen Benutzer oder eine Gruppe
+scm.namespace.set-role.namespace = Namespace
+
+scm.namespace.remove-permissions.usage.description.0 = Entzieht einem Benutzer oder einer Gruppe einzelne Berechtigungen
+scm.namespace.remove-permissions.namespace = Namespace
+
+scm.namespace.clear-permissions.usage.description.0 = Entzieht einem Benutzer oder einer Gruppe alle Berechtigungen
+scm.namespace.clear-permissions.namespace = Namespace
 
 ## User
 scm.user.username = Benutzername
